change model to only use Person object
